Added checking that external id is found before starting to delete existing areas

+++ leasing/importer/area.py
@@ -391,12 +391,14 @@ class AreaImporter(BaseImporter):
                     areas.update_or_create(defaults=other_data, **match_data)
                 except IntegrityError:  # There should only be one object per identifier...
                     ext_id = other_data.pop("external_id", "")
-                    # ...so we delete them all but spare the one with the correct external_id (if it happens to exist)
-                    Area.objects.filter(**match_data).exclude(
-                        external_id=ext_id
-                    ).delete()
-                    match_data["external_id"] = ext_id
-                    Area.objects.update_or_create(defaults=other_data, **match_data)
+                    # If external id exists, we can continue deleting data.
+                    if ext_id:
+                        # ...so we delete them all but spare the one with the correct external_id (if it happens to exist)
+                        Area.objects.filter(**match_data).exclude(
+                            external_id=ext_id
+                        ).delete()
+                        match_data["external_id"] = ext_id
+                        Area.objects.update_or_create(defaults=other_data, **match_data)
 
                 imported_identifiers.append(match_data["identifier"])
